Move serializer and deserializer implementation
