+++ dds/src/dds/topic_definition/type_support.rs
@@ -1,21 +1,23 @@
 use std::io::Read;
 
 use crate::{
-    serialized_payload::{
-        deserialize::CdrDeserialize,
-        deserializer::ClassicCdrDeserializer,
-        endianness::CdrEndianness,
-        parameter_list_deserialize::ParameterListDeserialize,
-        parameter_list_deserializer::ParameterListCdrDeserializer,
-        parameter_list_serialize::ParameterListSerialize,
-        parameter_list_serializer::{ParameterListCdrSerializer, ParameterListSerializer},
-        serialize::CdrSerialize,
-        serializer::ClassicCdrSerializer,
+    implementation::{
+        data_representation_builtin_endpoints::parameter_id_values::PID_SENTINEL,
+        payload_serializer_deserializer::{
+            cdr_deserializer::ClassicCdrDeserializer, cdr_serializer::ClassicCdrSerializer,
+            parameter_list_deserializer::ParameterListCdrDeserializer,
+            parameter_list_serializer::ParameterListCdrSerializer,
+        },
     },
-    implementation::data_representation_builtin_endpoints::parameter_id_values::PID_SENTINEL,
     infrastructure::{
         error::{DdsError, DdsResult},
         instance::InstanceHandle,
+    },
+    serialized_payload::{
+        deserialize::CdrDeserialize, endianness::CdrEndianness,
+        parameter_list_deserialize::ParameterListDeserialize,
+        parameter_list_serialize::ParameterListSerialize,
+        parameter_list_serializer::ParameterListSerializer, serialize::CdrSerialize,
     },
 };
