+++ dds/src/implementation/mod.rs
@@ -2,6 +2,7 @@ pub mod actors;
 pub mod configuration;
 pub mod data_representation_builtin_endpoints;
 pub mod data_representation_inline_qos;
+pub mod payload_serializer_deserializer;
 pub mod rtps;
 pub mod rtps_udp_psm;
 pub mod utils;
